[feat] Change to job assignment policy to worker threads
1. Worker threads use adaptive spinning to check for new jobs in
their queues, meaning they will sleep if a job is not found within a
time limit

2. When assigning a new job, the spinning thread's worker group array
is iterated in order to pick a worker with less than max_used_slots
tasks that is preferably already active

NOTE: Protocol operations still use the old policy, where a single
connection's jobs are assigned to a single worker. This is to make sure
that these operations are executed sequentially, in the order that they
arrive.