[feat][refactor] Clients check WC status of posted RDMA writes
1. kreon_rdma_client.c API calls now use the new
client_send_rdma_message function which checks the work completion
status of the posted RDMA Write. This requires a newly introduced
mechanism where each message can be associated with a context and a
semaphore. The caller can wait on the semaphore for the work completion
and this semaphore is posted by the poller thread (poll_cq &
on_completion_client functions). This mechanism is one of the
requirements for implementing failure detection in our client API

2. Refactored kreon_rdma_client.c API functions to use
_kc_wait_for_message_reply instead of duplicating the code in the
function bodies