[fix] Segment barrier STAGE during FLUSH to REPLICA
Primary server of a region sends a flush command to the replicas
without checking if the previous replication requests have completed.
This results in possible corruption of the replica's region data.
This commit adds a counter per segment counting the actual bytes
replicated in the remote segments
Second, Primary chooses the appropriate segment to replicate in
a Backup region based on the log_segment_full_event reported by the
storage engine. Checks for chosing the segment are ambiguous since this
offset should be >= from start and <= from end. Since
log_segment_full_event cannot occur for start this commit replaces this
with > start and <= end.