[fix] Data race in region's master segment metadata
System fails during replication because of the following data race.
Tasks that are in the replicate stage look up replicas segment
metadata to find the proper location to write their value. This
metadata are written by tasks that finish their flush
commands with the replica. This results in choosing wrong segment
to replicate. As a result the system enters a deadlock because
tasks that are in the SEGMENT_BARRIER stage cannot reach the
condition where all segments bytes are replicated and proceed.
We fix this by adding a lamport counter lock in each segment
metadata.
Also, there is a bug in searching the pools for available task
buffers resulting in an endless loop. We also fix this.