[feat][fix] Sleep/wake mechanism now requires zero locking
The previous sleep/wake mechanism for worker threads required that the
spinning thread acquires a lock on each task scheduling desicion and
also had a concurrency bug because of the lock being released by the
worker thread before the call to sem_wait.

This new mechanism delegates all decisions to the spinning thread, with
the worker only checking his status to determine if they should be
spinning or sleeping, and thus not requiring any locks.