[feat][refactor] Add categories for key-value pairs
* It was necessary to change the way we picked the category a
key-value pair belongs. Previously we calculated the key-value category
in every compaction. This would result in heavy I/O traffic for
workloads that stored key-value pairs in the log since we would read
every key-value pair to determine its size. Now we use the slot array to
store the category of the key-value when inserting the kv in L0 thus
avoiding random I/Os to read the size of the kv and determining the
category of the kv during compaction.
* In this commit compactions are not supported yet.