[fix] Replace in cursor pos_in_segment with offt_in_segment
We used pos_in_segment % SEGMENT_SIZE to calculate remaining
bytes in segment. pos_in_segment is a full virtual address
for which there is no guarantee that is SEGMENT aligned. This
caused failure during recovery. off_in_segment represents the
relative offset of the cursor in the current segment.