[refactor][fix] Change lookup_in_tree to return offset in device for gc
In the previous implementation of GC we returned the actual data of the
KV and more specifically the value to the GC thread. This was
unnecessary and resulted in read I/Os. With this commit we return the
offset to the KV without reading the KV. Now GC compares the device
offset to understand if the KV is replaced by an update/delete.