[fix] Make db_close wait for compaction threads
When a region was closing, it notified compactions threads to stop.
A compaction thread then only stopped inserting keys into the next level
but continued with persisting log metadata that were invalid resulting
in data loss. To solve this db_close will wait for compaction threads to
finish. This has a downside that in case a large compaction is triggered
the latency will be high for db_close. In the future probably we will
have to make compaction threads revert the modified state to reduce high
latency in db_close in case of large compactions.