[refactor][fix] Multiple fixes described in the body
* Fixes the translation of the medium KV category when moving medium KVs
from the log to in place. More specifically, we counted the medium KVs
as KV separated when checking for the size that will need in the leaf.
This could result in leaf overflow and potential corruption.
* Introduces {first,last}_segment_btree_level_offt to hold the device
offsets for the first and last segments for each level of the b-tree
that each LSM level contains. More specifically, we construct each level
in a bottom-up fashion as described in the paper. To make space
reclamation easier we link the segments for each level of the b-tree.
Before introducing {first,last}_segment_btree_level_offt we were losing
the first segment thus resulting in memory leaks in the device.
* Introduces compaction_daemon.h to host private structs and functions
for compaction_daemon.c. This header is private and should not be
included in other headers in the project.
* Also, I changed the names for several variables to describe the
context better.