[refactor] Use -pedantic -pedantic-errors in compiler, refactor index
The storage format for the B+-tree index nodes used a flexible array
that was redundant. This commit activates -pedantic and -pedantic-errors
as compilers flags and made the compiler unable to build the project
when using the flexible array. Also, this commit makes -Werror optional
for local builds. Also, this makes it easier to package the library in
systems with different compilers and distributions.