[fix] Fix race between YCSB clients and main thread
When the main YCSB thread spawned the threads for the load or the run
phase it could exit before the spawned threads returned. The problem in
this scenario was that the spawned threads had references in the stack
of the main thread resulting in use-after-free problems reported by the
address sanitizer.