[fix] Keep soft state about the medium log segments touched
Adds a soft-state hash table which keep track of the segments of medium
log touched during tranfer from log to in-place. This covers
cornercases where due to updates we do not reference the first chunck
of a segment of the medium log. We use this to calculate the max
segment id and its corresponding device offset from where we can
start the trim operation of the medium log.