[fix] Remove excess I/O traffic on flush of region operation log
Previously, the entire tail log chunk was flushed to the device,
regardless of the amount of data written. This could be inefficient in
scenarios with frequent sync commands where only a small portion of the
log tail chunk has been modified. This commit enhances the sync
operation by checking and writing only the dirty portion of the tail log
chunk to the device. This optimization minimizes the amount of data
written during sync, improving the performance and efficiency of the
operation.