[feat] Optimize level_enter_as_reader protocol
This version uses LEVEL_ENTRY_POINTS (default 8) instead of a single, to
enter a device level (L0 is intact) as a reader. Each reader decides
which entry point to use based either (a)on its rtsdc register value for
x86_64 arch or (b) pthread_self() function. level_enter_as_writer, which
is called infrequently at the end of a compaction, acquires each lock of
each LEVEL_ENTRY_POINT to enter the level. This protocol removes
contention among readers and we observer up to 30% improvement in
throughput of get operations when all data fit in memory.