[feat][fix] Add ability in Parallax to use custom layouts
This commit introduces two major refactorings and a fix. Now, Parallax
can support different leaf and index node layouts for the device levels.
Inside device_level.h, we declare two APIs, leaf and index API. By
implementing these two APIs in a separate C file and offering functions
register(struct leaf_api *leaf_api) and register(struct index_api
*index_api), Parallax can use these APIs to access the leaf and index
nodes of the level. In the current phase, the application must call in
the constructor of the corresponding device level(s), its leaf, and
index node API. In upcoming commits, we will introduce a proper factory
pattern that changes dynamically the custom layouts. The default device
layouts are now dev_leaf.c and dev_index.c, which contains the slot
array.

The second refactoring is the introduction of the level_lookup function,
which is responsible for searching for a kv_splice inside a device
level.

Finally, we fixed an issue regarding the number of keys field in the
device level, which Parallax did not persist in the superblock of the
region.