[fix] Parse WAL log segments properly during recovery
This commit fixes the following faulty check on WAL recovery. When we
parse a log segment, we must check if its remaining space can fit a
kv_splice and its log sequence number (LSN). In the previous version, we
did not consider the LSN size. We solely relied on the minimum possible
size of the kv_splice.