[feat] Transform B+-tree index to B-link trees
This commit introduces a pointer at each leaf of the B+-tree index of
the device levels to their sibling leaf. As a result, the B+-tree index
of the device levels is now a B-link tree. We keep accounting for these
pointers at the level_write_cursor.c during compaction. We use this
property for scans to a) simplify their code base and b) make scans more
efficient since they do not need to traverse a stack with index nodes to
discover their sibling leaf. In particular, we extend the API of the
device_level.h to include functions to scan it.