[feat] Introduce SSTs and memory guards in Parallax
This commit is a work in progress [WIP]. It introduces the object of
SSTs and memory guards in Parallax. Parallax still performs
full-leveling compactions;  however, it organizes its levels in SST
units and keeps guards in memory for their location and fence keys. SSTs
are a fundamental addition to future features and compaction policies.
We have added the Manto test, which tests the internal API of the SSTs
and levels. SSTs still need to be integrated fully into the system.