[feat] Add level_compaction_scanner
This commit adds the level_compaction_scanner object. Parallax uses this
object to scan the level via direct IO and to bypass (not to pollute)
the mmap cache. Its primary use is for scanning device levels during
compaction. Manto adds the verify_comp_scanner function to test the
correctness of the level_compaction_scanner.