[WIP][refactor] Add recovery of guards
This commit adds the appropriate logic to a) free sets after a
compaction and b) recover memory guards on db_open. It also includes
refactoring, mainly moving the rul_cursor function to redo_undo_log.h.
Finally, it enhances the Manto test to check for the correctness of the
recovery of the memory guards.