[WIP] Add experimental feature to use buffered I/O
This commit introduces the -DUSE_DIRECT_IO=<0|1> option. By default,
Parallax utilizes DIRECT_IO (-DUSE_DIRECT_IO=1). Setting it to 0 enables
Parallax to use buffered I/O instead. It's important to note that this
feature is still experimental and doesn't offer recovery guarantees, as
it doesn't issue a flush() operation after pwrite() calls. The intent
behind this feature is to cater to Parallax setups in HPC environments
and temporary states.