[fix] Add last_leaf_offt info in SST
The previous version relied on zeroing the I/O buffer on SST creation to
detect when Parallax reached the last leaf of an SST. This commit adds
the last offset (within the SST) of the last leaf in the SST info.
Parallax adds last leaf offset info only in the SST flush operation. As
a result, we reduce CPU cycles for zeroing buffers.